requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pymysql>=1.1.0
lxml>=4.9.0
python-dotenv>=1.0.0
